Fixed checkstyle, PMD and SpotBugs issues
